refactor: extract AnalyticsTable data fetching into src/api/analytics
Moves fetchGroupAnalytics/fetchUserAnalytics and raw useQuery calls out
of the page component into a feature-sliced src/api/analytics/ module
with a shared query-key factory, per ADR-0001.
